+++ app/src/main/java/com/gkzxhn/helpout/entity/PublishRequestInfo.kt
@@ -3,5 +3,5 @@ package com.gkzxhn.helpout.entity
 
 data class PublishRequestInfo(
     var category: String? = "",
-    var reward: Int? = 0
+    var reward: Double? = 0.0
 )
